check if file exists

+++ build.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-rm example.py example_wrap.cxx *.o *.so
+if [ -f example.py ]; then
+	rm example.py example_wrap.cxx *.o *.so
+fi
 
 swig3.0 -c++ -python example.i
 g++ -fpic -c example.c example_wrap.cxx Word.cpp -I/usr/include/python2.7/
